Overhaul of VCF-based LD lookups
Modifications to LDFeatureContainer allow VFs to be lazy-loaded
or not loaded at all. For example, the REST LD endpoint only
wants variant names; we get these from the VCF during genotype
loading, and therefore saves a DB lookup to get VFs which can
be slow especially when the query is uncached.