RunSift: disconnect when done (disconnect_if_idle) to free up connections
RunSift only needs to be connected at the end of the job to store its
results. For the rest of the time it should be disconnected to free up
connections to the mySQL server. This allows the use of as many workers
as the farm allows without cluttering the server with connections.